+++ pkg/env-tests/env_tests_suite_test.go
@@ -5,27 +5,19 @@ package env_tests
 
 import (
 	"context"
-	"path/filepath"
 	"testing"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	resourceapi "k8s.io/api/resource/v1"
+
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
-	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
-	"k8s.io/client-go/util/flowcontrol"
 	featuregate "k8s.io/component-base/featuregate/testing"
 	"k8s.io/kubernetes/pkg/features"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-
-	kaiv1alpha2 "github.com/NVIDIA/KAI-scheduler/pkg/apis/scheduling/v1alpha2"
-	kaiv2 "github.com/NVIDIA/KAI-scheduler/pkg/apis/scheduling/v2"
-	kaiv2v2alpha2 "github.com/NVIDIA/KAI-scheduler/pkg/apis/scheduling/v2alpha2"
-	kueuev1alpha1 "sigs.k8s.io/kueue/apis/kueue/v1alpha1"
 )
 
 var (
@@ -45,40 +37,11 @@ var _ = BeforeSuite(func(ctx context.Context) {
 	logf.SetLogger(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)))
 
 	By("bootstrapping test environment")
-	testEnv = &envtest.Environment{
-		CRDDirectoryPaths: []string{
-			filepath.Join("..", "..", "deployments", "kai-scheduler", "crds"),
-			filepath.Join("..", "..", "deployments", "external-crds"),
-		},
-		ErrorIfCRDPathMissing: true,
-	}
-
-	testEnv.ControlPlane.GetAPIServer().Configure().Append("feature-gates", "DynamicResourceAllocation=true")
-	testEnv.ControlPlane.GetAPIServer().Configure().Append("runtime-config", "api/all=true")
 	featuregate.SetFeatureGateDuringTest(tt, utilfeature.DefaultFeatureGate,
 		features.DynamicResourceAllocation, true)
-
 	var err error
-	// cfg is defined in this file globally
-	cfg, err = testEnv.Start()
-	cfg.ContentType = "application/json"
-	Expect(err).NotTo(HaveOccurred())
-	Expect(cfg).NotTo(BeNil())
-
-	// Effectively disable rate limiting
-	cfg.RateLimiter = flowcontrol.NewFakeAlwaysRateLimiter()
-
-	// Add any scheme registration here if needed for your custom CRDs
-	kaiv2.AddToScheme(scheme.Scheme)
-	kaiv1alpha2.AddToScheme(scheme.Scheme)
-	kaiv2v2alpha2.AddToScheme(scheme.Scheme)
-	resourceapi.AddToScheme(scheme.Scheme)
-	kueuev1alpha1.AddToScheme(scheme.Scheme)
-	// +kubebuilder:scaffold:scheme
-
-	ctrlClient, err = client.New(cfg, client.Options{Scheme: scheme.Scheme})
-	Expect(err).NotTo(HaveOccurred())
-	Expect(ctrlClient).NotTo(BeNil())
+	cfg, ctrlClient, testEnv, err = SetupEnvTest(nil)
+	Expect(err).To(Succeed())
 })
 
 var _ = AfterSuite(func(ctx context.Context) {